Rename `pantheon-cache.php` to `pantheon-page-cache.php`
