+++ wp-content/mu-plugins/pantheon.php
@@ -10,7 +10,7 @@
 
 if ( isset( $_ENV['PANTHEON_ENVIRONMENT'] ) ) :
 
-require_once( 'pantheon/pantheon-cache.php' );
+require_once( 'pantheon/pantheon-page-cache.php' );
 require_once( 'pantheon/pantheon-updates.php' );
 
 endif; # Ensuring that this is on Pantheon
