+++ wp-content/mu-plugins/pantheon/pantheon-page-cache.php
